Data is saved individually, slight style tweaks, better error messages
I underestimated the amount of data csTimer saves. Saving it all as
a single object is too big and causing errors for a lot of people.
All sessions and properties are saved individually and retrieved
individually. The limits are 8MB per item and 100MB total. If someone
goes over that there's nothing I can do as far as I know.

I adjusted the popup html to add a footer and create a slightly
better layout.

If an error occurs when saving data the actually error message
will be displayed for better troubleshooting.